refactor: formatRelativeTime を src/lib に共通化し重複実装を削除
Agent-Logs-Url: https://github.com/42kharuya/MultiFace/sessions/df8325e1-85e7-40ff-9644-a3e1432b8b58
