+++ src/components/notifications/NotificationList.tsx
@@ -5,29 +5,7 @@ import { faceRepository } from "@/repositories/face-repository";
 import { activityRepository } from "@/repositories/activity-repository";
 import { type Notification } from "@/types/notification";
 import { type User } from "@/types/user";
-
-// ─── ヘルパー ──────────────────────────────────────────────────
-
-/** ISO 8601 文字列を「N分前」「N時間前」「N日前」に変換 */
-const formatRelativeTime = (isoString: string): string => {
-  const now = new Date();
-  const date = new Date(isoString);
-  const diffMs = now.getTime() - date.getTime();
-  const diffMin = Math.floor(diffMs / 60_000);
-  const diffHour = Math.floor(diffMin / 60);
-  const diffDay = Math.floor(diffHour / 24);
-
-  if (diffMin < 1) return "たった今";
-  if (diffMin < 60) return `${diffMin}分前`;
-  if (diffHour < 24) return `${diffHour}時間前`;
-  if (diffDay < 30) return `${diffDay}日前`;
-
-  return date.toLocaleDateString("ja-JP", {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-  });
-};
+import { formatRelativeTime } from "@/lib/format-relative-time";
 
 // ─── 通知アイテム ──────────────────────────────────────────────
